infoDAQ.C: the beamProperties file output is modified
expAsym.C:
if the deadtime is turned on, it will now always look for the corresponding text file to get the corrections.

asymFit.C:
the polFit file now has the fit status also as the ninth entry.

batchsub.sh:
the file is actually compiled with a debug case for every instance. Somehow inthis case, it does not crash as often



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5876 b9700a98-9937-0410-b777-f318d6eb5bff