Added wire and plane efficiency in QwDetectorInfo for use when analyzing simulated events.
This can be set for an entire plane with
  efficiency = 0.9
or for individual wires (wires 14, 15, 16)
  efficiency_list = 14:0.8 15:0.8 16:0.8

Hits will be randomly dropped with the expected probability.

This requires C++0x or C++11 (finally: string to double, finally: uniform random numbers).
Enable it by adding -std=c++0x or (starting with GCC 4.7) -std=c++11 in the CXXFLAGS in
GNUmakefile somewhere around line 520:
  CXXFLAGS += -std=c++0x

If you don't have C++0x or C++11 and you use a geometry file with efficiency lines, this
code will produce an error and exit.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5883 b9700a98-9937-0410-b777-f318d6eb5bff