Removed redundant linked lists of partial tracks from all tracking structures.
This commit changes slightly (3%) the number of tracks found.  The cause is that partial
tracks that were explicitly found to be invalid (duplicates of partial track with more
hits, too high chi^2 > 2000, too similar to other partial tracks in the same event in
terms of wires hits) were still included in matching from region 2 to region 3.  These
(questionable) tracks are now not included anymore.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5894 b9700a98-9937-0410-b777-f318d6eb5bff