stripMask.C: separated the strip mask function from infoDAQ.C, now that function deals with all the slow tree and config tree and can be called separately to figure out the details of what was going on during the beam. getEBeamLasCuts.C: added mean and RMS of beam current and laser distribution, to easily see the behaviour of laser and ebeam.
infoDAQ.C:
Now this macro does not influence the execution of any other macro in the package. It just indepenently goes through the slow and config tree and writes down several relevant beam parameters to file for review while analysis. Strip mask process is removed to another small macro as stripMask.C
The energy was being written after being fixed to 1.159, repositioned it so that I get on file the true energy.
The file which has beam energy information also has an appropriate header to identify the constituents. energy is preceeded by run# too.
changed the method of determining if a beam parameter (IHWP etc) changed during the run, I use GetMaximum and GetMinimum and if they not same, i get it to enter that into a file with suffix runRemark.txt 
currently check on beam current before filling the energy histogram
The two created chains are deleted at the end.
We've added Fast feedback on position and energy lock; cComptScintRateNorm, dipole set and readback current.
The error is written using a TString, so it writes as a tab separated string all the parameters that has changed during the run. 

expAsym.C:
Now when the getEBeamLasCuts.C macro is executed, the beamMax is also checked and if it is lesser than beamOnLimit, the execution of expAsym is exited immediately. 
the position of 'GetEntry(i)' was such that the lasPow variable was not using the value for the current quartet. This was a mistake, I have moved the position appropriately
deleting all the TChain formed with a 'new', this could be the reason for the hung terminal at the end of execution. 
removed the call of infoDAQ.C, and added the call of stripmask() which is used by evaluateAsym.C 

rhoToX.C:
writing the fileoutput into the run# directory as requested by G.

asymFit.C:
exits execution if infoDAQ returns any negative number not just '-1'
now saves the bgdAsym.png file to disk, like other files

edetLC.C:
writes a file (WhyNotExecuted.txt) with the returned values from the executed function, indicating why possibly a run was not executed properly, as in, which macro failed



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5900 b9700a98-9937-0410-b777-f318d6eb5bff