Globally changed the variable name 'webDirectory' into 'www', for ease
asymFit.C:
added grid to bgdAsym plot; made all the plots of same dimension. Set limits on asym plot to show from 0 to 65 on x-axis


writeToFile.C:
removed the complicated method of adding a new line. this is possible since I can read the file in a smarter way. This was needed otherwise, the last strip(#64) was not being read and plotted in the version previous to this.

expAsym.C:
now using all good laser cycles or not, is a Boolean flag to be set in comptonRunConstants.h. This boolean is also written to a file in edetLC.C

evaluateAsym.C:
decoupled the calculation of background asymmetry from the regular asymmetry evaluation.
earlier if the asymErSqr for any lasCyc was non-positive, the macro quitted, now it checks if that strip is a masked strip, and if it is, that strip is skipped by 'continue' while using the lasCyc still




git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5903 b9700a98-9937-0410-b777-f318d6eb5bff