evalBgdAsym.C: trivial change of printing more relevant numbers in case of debugging
edetLC.C: 
removes the lines explicitly loading the macros, this won't be needed since we already 'include' it

infoDAQ.C:
Energy value was not properly set for cases when the energy measurement returned zero(the max and min were within 1MeV in this case, hence the macro was getting fooled).
Changed format of beamProperties file, now it has energy and RMS along with HWP readbacks; making it 10 entries each line
Changed format of dipole current output file, it also has just one line of values, 11 entries now

expAsym.C:
Removed the bpm related plots and have placed that into the studyRun.C macro which is not yet in svn. 
repositioned the command to create a new directory

asymFit.C:
it was missing the call to infoDAQ.C. without this call, the asymFit.C cannot be run independently, though it will work fine in the whole package

rhoToX.C:
trivial, added a statement for debug



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5915 b9700a98-9937-0410-b777-f318d6eb5bff