infoDAQ.C: the dipole current (set and readback) values are commented out. I realize its reasonable for them to change and its needless to monitor it.
expAsym.C
updated the noiseSubtraction to use a beam off file based on which run number is being analyzied. This was done particularly for the MCM runs.

evalBgdAsym.C & evaluateAsym.C:
changed the counts variable to a Double because I return a modified set of counts which will not be an integer when they are returned from the bgdCorrect function


asymFit.C:
slightly moved the position of paveText where the pol and fit values show up

some more trivial changes in other files



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5927 b9700a98-9937-0410-b777-f318d6eb5bff