used the r<#> structure rather than run_<#> structure to write files in all the macros that dealt with file writing reading. the directory created was changed to r<#> instead of run_<#> in comptonRunConstants.h and infoDAQ.C as well
in expAsym.C, qNormVariables.C and header file:
not using the current anymore, the charge is the only variable tracked through the laser cycles now.
Moved the kDeadTime module that determined the correction factors entirely before the nCycle loop. The factors are not affected by individual strip rate so there was no need for their calculation inside the loop.

using write3CfileArray.C and read3Cfile.C to read and write files . am currently not using the writeTofile.C macro; hence the fortran format output will not be printed if this version of svn revision is used.

all the txt files are now being written into a directory txt/ while the image files are written into www/ as usual; thisis for the ease of showing it on the web. 

the deadtime correction factors are being determined now in a separte macro named corrDeadtime.C

the studyRun.C is a macro to be used independently for checking beam and laser for that run



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5937 b9700a98-9937-0410-b777-f318d6eb5bff