This comment is for some scripts for ValerianROOT - to use in the analysis of the ValerianROOT files.  I will lay out all the details below, but in a nutshell, I have added and old traditional ROOT macro method in case one would like to write one that way, this is the Q2 old way script.  I added a new and wonderful Q^2 vs Run number script which uses the structures that I created and used in ValerianROOT for this reason.  I added a Treelines vs Run script also.  And since all scripts will read in a list of runs, which will have to get organized and read into a vector I created a script to do just this, which then can be just included in all ValerianROOT scripts.  Also in order to automatically set a good Y-axis range a less than but not zero value must be found - this is also done in a different script since all scripts will need this.
So in detail:

Q2_vs_Run_Number_typical_ROOT_method.C - This file graphs the Q2 vs Run number.
This file uses the traditional ROOT method to write a root macro.  So using
TBranches, TLeafs getting those values, etc.  This is a fine way to go about this,
but is not the best and most efficient since ValerianROOT was made to make things
a easier and calming and relaxing.

Q2_vs_Run.C - This script graphs the Q2 values vs run number.  It does the same
thing as the Q2_vs_Run_Number_Old_ROOT_method.C script *but* it uses and maybe
abuses the structures and ways I specifically set up ValerianROOT in the making
of ValerianROOT and the ValerianROOT tree.

TreeMatch_vs_Run.C - this script takes all the slope and Chi^2 for the tracks,
partial tracks and treelines.  The plots are for combinations of the partial
tracks and tracks, and then then treelines separately as they don't have the same info.

less_than_but_non_zero.h - This script is a little function that calculates
the smallest number in a vector that is not zero.  This is needed for setting
the lower value of range on the Y-Axis.  The upper range can be set with a simple
maximum.  If only everything was that easy ;-) If you are not graphing variable vs run
and instead are graphing variable vs variable this will need to be used on both the
X-Axis and Y-Axis to find the minimum value of the range.

Read_in_Run_Numbers.h - This function takes a list of run numbers, reads them in stores
them in a vector and organizes them from smallest to largest.  This script is needed by
all scripts since they will all need a list of run numbers.  Even if the scripts don't
graph variable vs run - since variable vs variable over certain run numbers still need
those run numbers so that the numbers for the variables are stored in such a way that
the corresponding entries are for the same run number.  



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5942 b9700a98-9937-0410-b777-f318d6eb5bff