Initial commit of ScalerCount class. This class processes a given rootfile and performs a basic scaler analysis.
At the moment, only the "raw" data should be trusted.
It also plots the ADC spectra for each PMT, and also all bars.

This is still a work in progress - one must still put a discription
into the histogram_adcs.cpp file. This should be implemented at
a later date, but at the moment is unnecessary.

Recommended compilation:
g++ -Wall `root-config --libs --cflags` \
    -Iinclude src/ScalerCount.cpp histogram_adcs.cpp -o adchisto

    To run:
    ./adchisto <filename> <low event> <high event>



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5944 b9700a98-9937-0410-b777-f318d6eb5bff