comptonRunConstants.h: added a variable beamTripLimit to be maintained at 78% of the beamMax to identify beam trips. some variables were changed to Float from Double
getEBeamLasCuts.C:
instead of cutting on the fixed 20uA, the beam trip is identified by 78% of the maximum.

expAsym.C:
in the last revision, the method of handling 'nthBeamTrip' was throwing out of bound error for the last beam trip in some cases (runs). This is fixed in this version.
Added a cut on helicity correlated beam position difference being larger than 0.05 in its own unit (nm). 

studyRun.C:
the output file is now named with a prefix of "study_"

anything else is a trivial change which is getting pushed with other files in this directory



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5945 b9700a98-9937-0410-b777-f318d6eb5bff