comptonRunConstants.h: added variables for tNorm printing for each helicity
getEBeamLasCuts.C:
now the beam drop is identified by a 50% of the beamMax instead of the flat 20uA limit. This helps avoid confusion when the beam rises and falls several times while trying to ramp up, when it stays to low current during each of its failed attempts. 

batchsub.sh: trivial

expAsym.C: now calculates the charge even for the laser cycles which had a beam trip. It also calls writeToFile.C now to write the fortran version of the output

writeToFile.C: it only writes the fortran version of the output, all other outputs are now written using write3CfileArray.C macro



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5956 b9700a98-9937-0410-b777-f318d6eb5bff