it wasn't appropriate to use the qNormVariables.C macro to get time normalized values, this wasn't coming out correctly. Fixed in this version. Most of the ver003 analysis will use this output.
The comment for the previous commit which got missed is here:
added a calculation output of variables about qNormH0L* counts, earlier on a test basis, it was done only for H1 type data. I've also added their tNorm counterparts.
Also the file name output named as qNormH0L0 was wrong, it contained qNormH1L0 counts.
The part of macro which was under the 'laserCyclePrint' boolean, has been moved to another place holder macro 'lasCycOut.C'; this part is not used for active analysis anymore and the standalone macro can be made to work when I want the quantities to be printed every laser cycle.
The cut on the helicity correlated difference was (by mistake) set to 0.05 mm, this has been fixed to 0.00005. The 0.00005 is in m (=50nm). 

rhoToX.C:
the return value of infoDAQ is checked and if not executed, it is called from rhoToX.C as well. This is another macro that uses energy, so thinking of modular coding, that too should have that called. 
the 'daqCheck' was renamed to 'retInfoDAQ' across all files that use it; to make the name more intuitive

infoDAQ.C and corrDeadtime.C:
removed the reference of 'dataType', hence changing the way it is called in every file that calls it



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5964 b9700a98-9937-0410-b777-f318d6eb5bff