the edetLC.C, expAsym.C and evaluateAsym.C together make sure that they return different negative values to identify which type of failure caused the macro to exit.
evaluateAsym.C:
An active variable checkAsym is maintained to check if the sign of asymmetry has changed midway during the run.


comptonRunConstants.h updated to allow the above changes to happen

batchsub_header1.xml updated to require only 1GB of memory, instead of requesting 2GB of memory. This will increase the no.of these jobs that can run at a time on farm.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5972 b9700a98-9937-0410-b777-f318d6eb5bff