corrDeadtime.C: if the given run# is not found in the appropriate file, then it exits, not just from the loop but from the macro execution. Appropriate changes made in expAsym.C to correspondingly set the correction factors to 1.0, so that in such a case, no correction is applied.
asymFit.C:
now the pol% and related numbers are written to 3rd decimal place. 

getEBeamLasCuts.C:
fixed the repetetive prints; also prints the pattern_number, along with entry #
The laser off time continuity has been chnaged to 13s from earlier 10 seconds. The smallest laser off with which I remember running was 15s.

expAsym.C:
writes out the run#, nCycle# and the counts into a file, whenever the evaluateAsym.C exits with -3 (detection of sign change of asymmetry); changed sequence of print statements;
turnsOn only the branches related to the plane being analyzed;
explicitly sets the counts variable to zero in the beginning of each laser cycle;
write out a fortran version of the output for every laser cycle, all in a single file.
The noise subtraction part of code moved to evaluateAsym.C. The beam position cut made more generous at 0.05 for all 3 bpms

comptonRunConstants.h:
increased the minimum no.of entries for continuous laser off to be 3120 instead of 2400.

writeToFile.C > writeFortOut.C:
the file name is changed and also the operation modified to output an counts per laser cycle, so that it can be later used without having to deal with the rootfile. 

evaluateAsym.C:
noise correction is now applied in evaluateAsym, hence #of quartets also passed in to the function. If the asymmetry in strip 39 changes sign, this asymmetry is not added to the running sum for weighted mean and the execution is exited. this condition will need to be relaxed for MCM runs.

studyRun.C:
histograms the laser power for the run also

batchsub.sh:
loads in corrected file writeFortOut.C and tNormVariables.C which were missing from last svn commit of this file




git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5976 b9700a98-9937-0410-b777-f318d6eb5bff