Weird bug fixed that only affects Anna and Josh :-)
In QwTracking there is an explicity Clear of all command line and config file options,
to force a reparsing of the options with the correct run numbers.  For some systems,
the farm being one of them (boost 1.41.00 in case anyone cares), this clears all options
and doesn't reparse them correctly.  For other systems (all of my systems) the options
don't get cleared.

I modified the code so it doesn't clear the options, but just forces a reparsing.  This
seems to fix things on the farm.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5977 b9700a98-9937-0410-b777-f318d6eb5bff