Edited this script that read in the files off the website.  These are the handful of runs that didn't have the same form as a file name as the others did. I changed that in the changes in the autoplots, and this was a change that followed on it.
Now all the scripts in this directory are all of the same form as
the file names are the same. In the future if there is added info
that needs to get added to the website and therefore ValerianROOT
as long as they output text file to the webpage is of the same
form, the modifications that would be need to added that to
ValerianROOT are easier since one doesn't have to worry they
grab a file that has a file name of a different form.  Since there
was only a handful of these it was easiest to do this.  This gets
rid of the chance of getting a file of a different form in the file you
are writing due to a copy issue.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5982 b9700a98-9937-0410-b777-f318d6eb5bff