combFiles.C: adds polarization and its error from 3 different version of analysis. Also the trigger type, scintillator rate and beam current information are added.
slug.C:
combines the runs in groups of slug. The weighted mean of polarization is continued until the sign of polarization changes. This slug combination does not track the actual HWP 1/2 or the wien states, just the polarization sign.

vectorMinMax.C:
gets the maximum and minimum of a vector used for plotting

allRunPol_3v.C:
analyzes the file generated by combFiles.C; the file has to be adapted if a corresponding change is done in the combFiles.C. 

cluster.C:
even if the group does not get enough number of runs to form a cluster as big as required by the 'group'; the remaining runs are still clustered. 

wholeRun2.C:
this macro is intended to be the one which will create most of the plots from run2 which do not pertain to cross analysis comparisons. 

simpleAvg.C:
the condition of the group being as large as 'group' is relaxed, so that values at the end which are smaller than the 'group' in which it is clustered are also averaged.

erWeightedMean.C
calculates teh errorweighted mean for a given input file

combRates_qNorm.C:
on lines of combFiles.C, this macro combines the rates of different versions of analysis for easy comparison.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5990 b9700a98-9937-0410-b777-f318d6eb5bff