Allow setting the minimum and maximum step for the RKF45 numerical integration method when tracking the electron through the magnetic field.  Defaults unchanged at 1 mm and 10 cm.
Allow setting the tolerance on the allowable truncation error when determining the new
step size in adaptive step numerical integration (RKF45, the default method).

Store number of iterations for fixed step RK4 and adaptive step RKF45 integration.


BUG FIX: Adaptive step was being determined incorrectly.  The step size is supposed to
be rescaled from h to h * sqrt(sqrt( tolerance * h / (2 * |Delta r|) )), but the multi-
plication with h was missing...  The effect of this, however, is very small.

Before this was fixed, the average number of steps was 165 over 577+250 cm, or a step
size of on average 5 cm.  The RMS difference between the 1 cm fixed step RK4 and the
adaptive step RKF45 was 7 micrometer around an average consistent with zero.

After this fix, the average number of steps is now 90 over 577+250 cm, or a step size
of on average 9 cm.  The RMS difference between the fixed step and adaptive step method
is now 17 micrometer with an average consistent with zero.  The marginally reduced
agreement is in line with the reduction in number of steps.


What this also points out is that the numerical integration is likely implemented
correctly, that we are not pegged at large or small step size, and that different
integration methods are in agreement at a level much better than what we need.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@6010 b9700a98-9937-0410-b777-f318d6eb5bff