In terms of analysis this is complete with that was on the list of additions.
We use the magnetic field extracted from the dipole current. The MCM runs had a field of -1000 Gauss-cm due to the trim coils, that correction is also included in the calculation. This is done via the ValueLookup.C which has a look up table of the value hashed from the string read back of the dipole current. This macro takes into account dipole current in the range 90-110 uA. 

I'm masking strip 60, and that makes no difference in the statistics or the result. We use the noise from run 25286 for all the noise corrections in the MCM runs. 

We find the warnings pertaining to dipole current needless hence commented out in infoDAQ.C. 

mcmRuns.C can plot additional plots for comparison and evolution of polarization during MCM. 

Other changes are generic and trivial

We use 


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@6013 b9700a98-9937-0410-b777-f318d6eb5bff