in MCM directory & edetHelLC directory:
determineNoise.C:
writes the noise rate into a file named NoiseRateP.txt along with the regular prefix.

added ValueLookup.C to find the dipole current 
added a new file writeFortFile.C:
unlike writeFortOut.C which writes out the counts every laser cycle, this macro will output the counts at the end of a run, to easily compare uncorrected counts between runs.

comptonRunConstants.h:
the qIgnored* moved to expAsym.C as a local variable and is passed in as parameter to 

expAsym.C:
on detecting a change in sign of asymmetry, the macro does not exit the execution entirely; but exits the nCycle loop and tries to extract polarization from the laser cycles upto that where the change was detected. The cycle where this was detected is not used.

Other than the above, propagated the changes made in MCM directory for iterative rhoToX determination until the new CE obtained is within 0.1 of the previous CE.




git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@6023 b9700a98-9937-0410-b777-f318d6eb5bff