New values for Region 2 (HDC) time-zero's (t_0) for Run 2.
 These are based on my empirical optimization study:
     http://dilbert.physics.wm.edu/Physics/107

  The file ".12164-30000.map is the default for non-reverse
runs (where th same package was used for Region 2 and Region 3 in a given
octant). The others are for those cases where either Region 2 or Region
3 was in a "reversed" configuration. One (8668-8673) is for Run 1,
but we haven't checked the t0's for Run 1 yet, so this may be 
suspect. Need to check the default T0's for Run 1 as well.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@6026 b9700a98-9937-0410-b777-f318d6eb5bff