setup cernlib and setup root added to the batchsub_header2.xml
evalBgdAsym.C , evaluateAsym.C, writeFortOut.C, writeFortFile.C:
uses the correct time for laser on/off instead of the earlier method where I used to get time from no.of helicity/helRate.

weightedMean.C:
a modular function that just carries out the weighted mean calculation and is called multiple times by the main code. Earlier this macro was calculating the weighted mean of the individual quantities

ValueLookup.C:
In the output file, the actual magnetic field is also printed along with the dipole field

comptonRunConstants.h:
bpmDiff=0.4 (from 0.5)
lasOffCut=1182(equivalent of 1.1 in the raw laser power)

asymFit.C:
allowed the CE to vary up to 63, earlier it was limited to 62. few runs around 25419(not 24519) have the CE at strip 63.
In the case of fit failure, the parameters are released and the new initial CE is 0.5 strip behind the previously attempted CE.

stripMask.C:
added strip 60 to the list of masked strips, for now, this may have to be removed

expAsym.C:
added reading the 4mhz clock to determine the true time elapsed. The changes in weightedMean.C reflect here too.





git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@6039 b9700a98-9937-0410-b777-f318d6eb5bff