Split up tracking R3 HDC geometry description into simulation and data versions.
The simulation version is made to agree with the current default settings in the
geant4 simulation geometry description.  All octants are identical.  If changes
are made to the geant4 geometry, this geometry file needs to be updated as well.

WARNING: There are still some disagreements between geometry here and the geant4
simulation.  This is because of some missing messengers in geant4, I think.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@6054 b9700a98-9937-0410-b777-f318d6eb5bff