Merging changes in branch QwAnalysis_3.04 into trunk.
Because of restrictions on qweaksvn (version 1 of svn file structure, instead of version 3)
we are limited in doing a diff merge: differences are merged, history is lost.

  svn merge https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@HEAD https://qweaksvn.jlab.org/repos/QwAnalysis/branches/release_3.04@HEAD .

No conflicts.

Also, if you look at the revision log for 6107, you will notice that this largely overlaps
with the files that were added to the branch earlier.  That means no unexpected changes snuck in.

The qweak_hallc_beamline.5000-.map change is consistent, and allows for pre-Wien0 analysis.  That
had not been merged into trunk yet (or indepedently uploaded there).



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@6110 b9700a98-9937-0410-b777-f318d6eb5bff