This script adds on to the existing Q2_vs_Run.C macro. It adds on correcting for the beam position. Which allows to compare the Q2 values for muliple runs at the same beam positon, which we can't garrenntee for tracking runs. Exspecially in different run periods.
This beam position correction is done in an eliptical method. We have
2 well known beam posion relationships which have been well investigated
in both data and simulation and they agree in these cases. These two are
the Q2 and the scattering angle. 

In order to make this correction It uses a run for the "nominal"
beam position. "Nomial" is definded asthe run that's beam position is
used and all runs are "corrected" to this beam position.   
This is done for each pkg individually, and by using the angle values
 and then it is applied the Q^2 values.
 - Correct for nominal run to be in the nominal beam position
 - Beam position determined by the scattering angle value. Position_Scan_Q2_ScatteringAngle
 - The Q2 angel values are from tracking elog https://qweak.jlab.org/elog/Tracking/83
 - After getting the beam position then the Q2 is corrected to "nominal" beam
 position, which will be from the nominal run
 - The errors is not corrected for.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@6178 b9700a98-9937-0410-b777-f318d6eb5bff