This commit has the Minuit code added to QwTrackingTreeCombine.cc, as well as including the Minuit libraries in QwTrackingTreeCombine.h.  Currently the Minuit code rotates each HDC chamber around it's center.  This may be changed so that both chambers in a package rotate around the same point at a later date.
It also changes the HDC geometry file (qweak_HDC.5000-30000.geo) so that includes the variables detector_roll and detector_pitch, which are currently set to 0.  I should be commiting a geometry file soon that will have the detector_roll (xrot) and detector_pitch (yrot) set to the values to realign the US and DS chambers, according to the nov 2011 running period.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@6208 b9700a98-9937-0410-b777-f318d6eb5bff