Recently (i.e. yesterday) David made me realize again that the decision to optimize the momentum on the radial POSITION of the back partial track at the virtual matching plane before the VDC chambers is arbitrary.
This commit implements a generalization of the optimization variable in Newton's method,
and allows to pick the DIRECTION of the momentum vector at that matching plane as the
optimization variable.

Initial tests show that for individual events the difference between the two Newton's method
optimization strategies is randomly distributed with an RMS (by eye) of about 10 MeV on the
momentum of the track. The difference on the mean of the distribution is however a lot smaller
(less than 0.5 MeV on a sample of 1000 events).

Nevertheless, this should be studied in a bit more detail. I only looked at this for
simulated events where this is expected to not have an effect. Incorrectly implemented
geometry in the actual data may however have different impact on position than on direction.

How to use this? The old option --QwRayTracer.position_resolution (0.3 cm) has been replaced
by the options
--QwRayTracer.optim_variable (0) with 0 for position.Perp() and 1 for direction.Theta()
--QwRayTracer.optim_resolution (0.3 cm) sets the value for cut-off
The framework allows for other optimization variable, but I can't think of any that make
sense (and are not plagued by normalization issues, in case someone suggests to use the
quadratic sum of position.Perp() and direction.Theta()).

The default is an 0.3 cm resolution:
  --QwRayTracer.optim_variable 0 --QwRayTracer.optim_resolution 0.3

No one was using the old option and it is in none of our configuration files, so I didn't
think it was necessary to keep the old command line option. If someone really needs it, let
me know.

A cut-off value of 0.001 rad for direction.Theta() leads to the same number of iterations
of the Newton's method (so probably the same precision, but who knows). If you want to use
this, use the following (for a 1 mrad resolution):
  --QwRayTracer.optim_variable 1 --QwRayTracer.optim_resolution 0.001

The average number of iterations of Newton's method is between 2 and 3. The average number
of iterations for the RK method swimming through the magnetic field is 200. Keeping the
number of iterations in the Newton's method small keeps the processing cost low.

The position.Perp() optimization remains the default.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@6258 b9700a98-9937-0410-b777-f318d6eb5bff