Read and store crosstalking 'elements' (wires) for detectors (wire chambers).
In the main detector.map, add the following line to the respective section:

  cross = qweak_VDC.cross

and create a file qweak_VDC.cross (see below for note on filename):

  [VDC_pkg2_back_u]
  crosstalk = 114,115
  crosstalk = 120,121
  [VDC_pkg2_back_v]
  crosstalk = 104,105
  crosstalk = 110,111

This will (in the future) treat those wires as crosstalking in both directions.
Right now the info is stored as a map<int,int> in the QwDetectorInfo object of
the respective plane.

You must use a detector name as defined in the geo file. If there's a mistake
it will exit(-1), so don't do that.

You can define crosstalk channels for any detector. What it means is up to the
individual chambers to decide.

You can use the usual run number filename scheme for crosstalk that occurs only
in some runs and not in others. Naturally, simulation will NOT have crosstalk
without further work.

Only pairs are defined. If you have three or more elements that are crosstalking,
tough luck right now. If you want to try something clever like adding two lines
with different combinations, don't: it won't work.

Next up: cutting the worst chi2 of two crosstalking wires in the same VDC tree
lines, and repeating the chi2 fit. The underlying assumption is that the hit on
a crosstalk wire is real and belongs to one (and only one) of the two wires.
All information about any hit on the other wire is assumed lost.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@6342 b9700a98-9937-0410-b777-f318d6eb5bff