Don't you ever go naively subtracting unsigned ints again!!
This bug, in its day-long frustrating beauty, illustrates the importance of
taking a class or reading up on data types. In decoding the F1 TDC channels
for the HDC, one gets an unsigned integer value. Of course, because of how the
F1 TDCs function, one needs to take the difference between the channel and some
reference signal, or between the channel time and the trigger time. All of those
values are stored at unsigned ints.

In the code this was written as (with some variables shortened):
  abs(reference_trig_time - header_trig_time);
where both reference_trig_time and header_trig_time are unsigned ints.

Because of the way C, C++ and most (all?) typed language treat operations, the
value of (reference_trig_time - header_trig_time), i.e. a subtraction of two unsigned
ints, will also be an unsigned integer. It will always be positive. Sure, you can
take the absolute value after you're done, but it won't change a damn thing.

What was actually meant here was
  abs(int(reference_trig_time) - int(header_trig_time));
Now the subtraction will produce a regular signed int. We upconvert the unsigned int
to a signed int. (Note: even safer would be a long since not every unsigned int can
be converted to a signed int of the same bit length, but here I know that the values
are well accommodated in a regular signed int.)

Why is this important? If, as was the case for event 25 in run 15121, the values
are reference_trig_time = 20 and header_trig_time = 21, then you end up with an
incorrect difference of 4294967295. Certainly quite a difference from the value
1 that the programmer intended to have gotten.

This resulted in the data buffer being thrown out because of crate synchronization
issues (SYN), and the event being discarded altogether because of lack of hits in
the HDCs.

There's no telling how many times this happens (event 25 is only the second event
in run 15121 with hits in the HDC).

It is also not clear in how many other places in the code these subtractions happen.

So, just stay away from subtracting unsigned ints.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@6356 b9700a98-9937-0410-b777-f318d6eb5bff