Deleting the cross-talking VDC wires requires a bit more care than I originally put into it...
As originally implemented, the code deleted the wire with the worst residual of
two cross-talking wires. This worked well in testing with some arbitrary
non-cross-talking wires since there was ever only one hit in e.g. wire 113 and
one hit in wire 114. Now that we know the actual cross-talking wires, the
situation is different: if wire 113 cross-talks to wire 114, there are (sometimes)
two hits on wire 114. Deleting the first hit reduces the number of hits such that
deleting the second hit deletes the wrong hit. Chaos ensues.

As a work-around: 1) create a list of which cross-talking wires to delete. This list
can contain duplicates (wire 113 could be worse than both hits on wire 114, but
should still only be deleted once); 2) Do the equivalent of | sort | uniq using
std::vectors; 3) Delete the hits starting from the back of the hit list, taking
care not to compare unsigned integer with zero (see previous commit log).

Minor issue with this fix:
- If there is a hit on wire 113 and two on 114, it seems that there is some
  directionality to the cross-talking: hit on wire 113 gets registered on wire 114,
  but not the other way round. We could be smarter about it and only remove the
  worst hit between wires 113 and 114 that have the same drift distance.
  Anyway, probably only affects a fraction of a fraction of events...



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@6368 b9700a98-9937-0410-b777-f318d6eb5bff